remove graph traits
