+++ src/core/graph/mod.rs
@@ -1,3 +1,4 @@
+pub mod command;
 pub mod diff;
 pub mod directed_graph;
 pub mod graph;
