+++ src/core/history/history.rs
@@ -1,10 +1,13 @@
+use std::collections::HashMap;
+use std::rc::Rc;
+
+use core::fmt::Debug;
+use core::hash::Hash;
+
+use crate::core::graph::command::GraphCommand;
 use crate::core::graph::graph::*;
 use crate::core::history::hashlist::*;
 use crate::core::history::history::Ref::*;
-use core::fmt::Debug;
-use core::hash::Hash;
-use std::collections::HashMap;
-use std::rc::Rc;
 
 type Hashs = Rc<HashList>;
